更改了登录方式 更新了部分接口

+++ src/dao/RolesDao.java
@@ -10,34 +10,46 @@ import java.util.List;
  * @// TODO: 2018/5/5
  */
 public class RolesDao {
-    Connection connRole=DBUtil.getConnection();
+    Connection connRole=(new DBUtil()).getConnection();
     //添加一种权限
-    public void addRoles(Roles r)throws SQLException{
-        String sql="insert into tb_roles (role_id,role_name,permission) values(?,?,?)";
-        PreparedStatement pstmt=connRole.prepareStatement(sql);
-        pstmt.setInt(1,r.getRole_id());
-        pstmt.setString(2,r.getRole_name());
-        pstmt.setInt(3,r.getPermission());
-        pstmt.execute();
+    public boolean addRoles(Roles r)throws SQLException{
+        if(searchSingleRole(r.getRole_id())!=null) return false;
+        else {
+            String sql = "insert into tb_roles (role_id,role_name,permission) values(?,?,?)";
+            PreparedStatement pstmt = connRole.prepareStatement(sql);
+            pstmt.setInt(1, r.getRole_id());
+            pstmt.setString(2, r.getRole_name());
+            pstmt.setInt(3, r.getPermission());
+            pstmt.execute();
+            return true;
+        }
     }
     /******************更新某个权限***********************/
-    public void updateRoles(Roles r)throws SQLException{
-        String sql="update tb_roles set role_name=?,permission=? where role_id=?";
-        PreparedStatement pstmt=connRole.prepareStatement(sql);
-        pstmt.setString(1,r.getRole_name());
-        pstmt.setInt(2,r.getPermission());
-        pstmt.setInt(3,r.getRole_id());
-        pstmt.execute();
+    public boolean updateRoles(Roles r)throws SQLException{
+        if(searchSingleRole(r.getRole_id())==null) return false;
+        else {
+            String sql = "update tb_roles set role_name=?,permission=? where role_id=?";
+            PreparedStatement pstmt = connRole.prepareStatement(sql);
+            pstmt.setString(1, r.getRole_name());
+            pstmt.setInt(2, r.getPermission());
+            pstmt.setInt(3, r.getRole_id());
+            pstmt.execute();
+            return true;
+        }
     }
     /******************更新某个权限***********************/
 
 
     /******************删除某种权限***********************/
-    public void deleteRole(int role_id)throws SQLException{
-        String sql="delete from tb_roles where role_id=?";
-        PreparedStatement pstmt=connRole.prepareStatement(sql);
-        pstmt.setInt(1,role_id);
-        pstmt.execute();
+    public boolean  deleteRole(int role_id)throws SQLException{
+        if(searchSingleRole(role_id)==null) return false;
+        else {
+            String sql = "delete from tb_roles where role_id=?";
+            PreparedStatement pstmt = connRole.prepareStatement(sql);
+            pstmt.setInt(1, role_id);
+            pstmt.execute();
+            return true;
+        }
     }
     /******************删除某种权限***********************/
